requirements.txt
awscli==1.16.125
awsebcli==3.14.13
opencv-python==4.2.0.32
Pillow==7.0.0
beautifulsoup4==4.8.2
svglib==1.0.0
ffmpeg-python==0.2.0
Flask==0.12.2
Flask-Caching==1.3.3
Flask-Compress==1.4.0
Flask-Cors==3.0.3
Flask-RESTful==0.3.6
google==2.0.2
google-api-python-client==1.8.0
google_auth_oauthlib==0.4.1
google-auth-httplib2==0.0.3
google-auth==1.13.1
schedule==0.6.0
SQLAlchemy==1.2.1
tweepy==3.8.0
instapy_cli==0.0.12
mhmovie==2.8
moviepy==0.2.3.2
pymysql==0.9.3
facebook-sdk==3.1.0
selenium==3.141.0